Renamed from Uno to NodeMCU
Also edited the description in contained within the DCCpp_NodeMCU.ino file.
